# openapi.yml
openapi: 3.1.0
info:
  title: Camplx Campus Circular Marketplace API
  version: 1.0.0
  description: 'Canonical backend contract for CAMPLX (CampusSwap AI). Facilitates
    peer-to-peer campus reuse, multi-angle computer vision product inspection, evidence-backed
    appraisal pricing, no-payment reservations, campus want-board ("Need It"), e-waste
    collection optimization, and verified sustainability metrics.

    '
servers:
- url: http://localhost:8000/api/v1
  description: Local FastAPI Development Server (iOS Simulator / Web)
- url: http://10.0.2.2:8000/api/v1
  description: Android Emulator Gateway
- url: https://api.camplx.internal/api/v1
  description: Staging / Production Backend Gateway
...
paths:
...
  /listings/{listing_id}/reservations:
    parameters:
    - name: listing_id
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - Reservations
      operationId: getActiveListingReservation
      summary: Retrieve the active reservation for a listing
      description: Accessible by the listing seller or the reserving buyer. Returns
        404 if listing is unreserved.
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      responses:
        '200':
          description: Active reservation record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationRecord'
        '404':
          description: No active reservation exists for this listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Reservations
      operationId: createListingReservation
      summary: Place a no-payment pickup reservation on an active listing
      description: Atomic transition. Rejects if another student holds an active reservation.
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      responses:
        '201':
          description: Reservation confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationRecord'
        '409':
          description: Conflict (Item is already reserved or completed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservations/my:
    get:
      tags:
      - Reservations
      operationId: listMyReservations
      summary: Retrieve listings reserved by the current student
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      responses:
        '200':
          description: List of reservations with embedded listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MyReservationItem'
  /reservations/{reservation_id}/release:
    parameters:
    - name: reservation_id
      in: path
      required: true
      schema:
        type: string
    post:
      tags:
      - Reservations
      operationId: releaseReservation
      summary: Release an active reservation back to available status
      description: Either the reserving buyer or the seller may cancel the hold.
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      responses:
        '200':
          description: Reservation released; listing returned to active status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationRecord'
        '403':
          description: Forbidden (Actor is neither buyer nor seller)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservations/{reservation_id}/complete:
    parameters:
    - name: reservation_id
      in: path
      required: true
      schema:
        type: string
    post:
      tags:
      - Reservations
      operationId: completeReservation
      summary: Confirm physical handoff and mark item sold or donated
      description: Seller-only terminal action. Increments carbon telemetry and rewards
        points atomically.
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteReservationRequest'
      responses:
        '200':
          description: Transaction completed and impact telemetry attributed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompletedTransactionResponse'
        '403':
          description: Forbidden (Only the listing seller may complete the handoff)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /wishlist:
    get:
      tags:
      - Wishlist
      operationId: getMyWishlist
      summary: Retrieve listings saved by the current student
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      responses:
        '200':
          description: Array of saved listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListingSummary'
    post:
      tags:
      - Wishlist
      operationId: addToWishlist
      summary: Save a listing to wishlist
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - listing_id
              properties:
                listing_id:
                  type: string
      responses:
        '201':
          description: Item bookmarked
        '404':
          description: Listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wishlist/{listing_id}:
    parameters:
    - name: listing_id
      in: path
      required: true
      schema:
        type: string
    delete:
      tags:
      - Wishlist
      operationId: removeFromWishlist
      summary: Remove a listing from wishlist
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      responses:
        '204':
          description: Item removed from wishlist
...
  /notifications:
    get:
      tags:
      - Messaging & Notifications
      operationId: getMyNotifications
      summary: Retrieve in-app notifications feed
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      responses:
        '200':
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationItem'
  /notifications/{notification_id}/read:
    parameters:
    - name: notification_id
      in: path
      required: true
      schema:
        type: string
    post:
      tags:
      - Messaging & Notifications
      operationId: markNotificationAsRead
      summary: Mark notification as read
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      responses:
        '200':
          description: Notification marked as read
  /notifications/read-all:
    post:
      tags:
      - Messaging & Notifications
      operationId: markAllNotificationsAsRead
      summary: Mark all notifications as read
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      responses:
        '200':
          description: All notifications marked as read
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: PocketBase or standard JWT token.
    DemoUserHeader:
      type: apiKey
      in: header
      name: X-Camplx-Demo-User-Id
      description: 'Development and demo identity bypass header (e.g. a0000000-0000-0000-0000-000000000001).
        Strictly disallowed and ignored in production environments.

        '
  schemas:
    ItemCategoryEnum:
      type: string
      enum:
      - electronics
      - books
      - cycles
      - furniture
      - hostel_essentials
      - other
    ListingTypeEnum:
      type: string
      enum:
      - sell
      - donate
    ListingStatusEnum:
      type: string
      enum:
      - active
      - reserved
      - sold
      - donated
      - inactive
...
    ConditionGradeEnum:
      type: string
      enum:
      - Like New
      - Good
      - Fair
      - Needs Repair
...
    ReservationStatusEnum:
      type: string
      enum:
      - active
      - released
      - completed
...
    ErrorResponse:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          example: INVALID_STATE_TRANSITION
        message:
          type: string
          example: This listing is already reserved by another student.
        details:
          type: object
          additionalProperties: true
...
    ListingSummary:
      type: object
      required:
      - id
      - seller_id
      - title
      - category
      - condition
      - listing_type
      - status
      - primary_image_url
      - created_at
      properties:
        id:
          type: string
        seller_id:
          type: string
        title:
          type: string
        brand:
          type:
          - string
          - 'null'
          example: Casio
        model:
          type:
          - string
          - 'null'
          example: FX-991EX
        category:
          $ref: '#/components/schemas/ItemCategoryEnum'
        condition:
          $ref: '#/components/schemas/ConditionGradeEnum'
        listing_type:
          $ref: '#/components/schemas/ListingTypeEnum'
        asking_price:
          type:
          - number
          - 'null'
        status:
          $ref: '#/components/schemas/ListingStatusEnum'
        carbon_savings_g:
          type: integer
          example: 3500
        primary_image_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
...
    ReservationRecord:
      type: object
      required:
      - id
      - listing_id
      - buyer_id
      - buyer_name
      - status
      - created_at
      properties:
        id:
          type: string
          example: res-102938
        listing_id:
          type: string
          example: list-1
        buyer_id:
          type: string
          example: b0000000-0000-0000-0000-000000000002
        buyer_name:
          type: string
          example: Priya Nair
        status:
          $ref: '#/components/schemas/ReservationStatusEnum'
        created_at:
          type: string
          format: date-time
    MyReservationItem:
      type: object
      required:
      - reservation
      - listing
      properties:
        reservation:
          $ref: '#/components/schemas/ReservationRecord'
        listing:
          $ref: '#/components/schemas/ListingSummary'
    CompleteReservationRequest:
      type: object
      required:
      - outcome
      properties:
        outcome:
          type: string
          enum:
          - sold
          - donated
    CompletedTransactionResponse:
      type: object
      required:
      - reservation_id
      - listing_id
      - final_status
      - carbon_savings_g
      - points_awarded
      properties:
        reservation_id:
          type: string
        listing_id:
          type: string
        final_status:
          type: string
          enum:
          - sold
          - donated
        carbon_savings_g:
          type: integer
          example: 3500
        points_awarded:
          type: integer
          example: 50
...
    NotificationItem:
      type: object
      required:
      - id
      - type
      - actor_name
      - message
      - read
      - created_at
      properties:
        id:
          type: string
        type:
          type: string
          enum:
          - like
          - reservation
          - request
          - message
          - system
        actor_name:
          type: string
          example: Priya Nair
        message:
          type: string
          example: Priya Nair placed a pickup hold on your Casio Calculator.
        read:
          type: boolean
          example: false
        target_entity_type:
          type:
          - string
          - 'null'
          enum:
          - listing
          - reservation
          - need_request
          - chat
        target_entity_id:
          type:
          - string
          - 'null'
        created_at:
          type: string
          format: date-time